% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auth.R
\name{construct_query}
\alias{construct_query}
\title{Reorder a query and URI encode the parameter names and values}
\usage{
construct_query(query)
}
\description{
Reorder a query and URI encode the parameter names and values
}

