% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create_auth}
\alias{create_auth}
\title{Create the final signature to be added to the HTTP header as Authorization.}
\usage{
create_auth(request_method, headers, payload, uri, query, date_stamp,
  region_name, service_name, request_date)
}
\arguments{
\item{request_method}{the HTTP verb used for the request (GET/POST/PUT etc)}

\item{headers}{list of headers to be included in the request. Must include a
\code{host} header. See examples for correct format}

\item{payload}{the payload from the body of the HTTP/HTTPS request}

\item{uri}{}

\item{query}{}

\item{date_stamp}{request date in the form YYYYMMDD; defaults to the current
date.}

\item{region_name}{name of the AWS region being targeted, e.g. 'eu-west-1'}

\item{service_name}{name of the AWS service being targeted, e.g. 'ec2'}

\item{request_date}{string containing the date and time of the request,
matching the value used in previous steps, in the form YYYYMMDDTHHMMSSZ}
}
\description{
This is the final step in the authorization procedure, where the three tasks
are put together to create the authorization value.
}
\examples{
headers <- list(
   'content-type' = 'application/x-www-form-urlencoded; charset=utf-8',
   'x-amz-date'   = '20110909T233600Z',
   'host'         = 'iam.amazonaws.com')
create_auth('POST', headers, 'Action=ListUsers&Version=2010-05-08', '/',
                 '', '20110909', 'us-east-1', 'iam', '20110909T233600Z')
create_auth('GET',
                 list('Date'='Mon, 09 Sep 2011 23:36:00 GMT','Host'='host.foo.com'),
                 '', '/', 'foo=Zoo&foo=aha', '20110909', 'us-east-1', 'host',
                 '20110909T233600Z')
}

