% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auth.R
\name{create_credential_scope}
\alias{create_credential_scope}
\title{Create the credential scope string.}
\usage{
create_credential_scope(region, service, date_stamp)
}
\arguments{
\item{region}{region being targeted}

\item{service}{being targeted}

\item{date_stamp}{date in the form YYYYMMDD - must match that used in other
steps}
}
\description{
Helper function for concatenating strings into the right format for the
credential scope value.
}
\examples{
create_credential_scope('20110909', 'us-east-1', 'iam')
}

