% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auth.R
\name{create_request}
\alias{create_request}
\title{Create a canonical request and hashed canonical request.}
\usage{
create_request(request_method, headers, payload, uri = "/", query = "")
}
\arguments{
\item{request_method}{the HTTP verb used for the request (GET/POST/PUT etc)}

\item{headers}{list of headers to be included in the request. Must include a
\code{host} header. See examples for correct format}

\item{payload}{the payload from the body of the HTTP/HTTPS request}

\item{uri}{the absolute path component of the uri}

\item{query}{the query string of the request. May be empty if the query is in
the payload instead (the default)}
}
\description{
This function puts together an http request into a standardised (canonical) form,
to ensure that the signature calculated by AWS when it receives the request
matches the one calculated by us. This is the equivalent of Task 1 in the AWS
API Signature Version 4 signing process
(http://docs.aws.amazon.com/general/latest/gr/sigv4-create-canonical-request.html).
}
\examples{
headers <- list(
   'content-type' = 'application/x-www-form-urlencoded; charset=utf-8',
   'x-amz-date'   = '20110909T233600Z',
   'host'         = 'iam.amazonaws.com')
create_request('POST', headers, 'Action=ListUsers&Version=2010-05-08', '/', '')
}

