% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auth.R
\name{create_string_to_sign}
\alias{create_string_to_sign}
\title{Create a string to sign.}
\usage{
create_string_to_sign(full_canonical_request, credential_scope, request_time)
}
\arguments{
\item{credential_scope}{credential_scope string calculated by the function
of the same name}

\item{request_time}{string containing the date and time of the request,
matching the value used in previous steps, in the form YYYYMMDDTHHMMSSZ}

\item{hashed_canonical_request}{hashed canonical request passed on from
create_request}
}
\description{
This function is the equivalent of Task 2 in the AWS API Signature Version 4
signing process
(http://docs.aws.amazon.com/general/latest/gr/sigv4-create-string-to-sign.html).
It currently only uses SHA256; this can be easily changed in future if
necessary.
}
\examples{
create_string_to_sign('3511de7e95d28ecd39e9513b642aee07e54f4941150d8df8bf94b328ef7e55e2',
               '20110909/us-east-1/iam/aws4_request\\n',
               '20110909T233600Z\\n')
}

