% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auth.R
\name{create_auth}
\alias{create_auth}
\title{Create the final signature to be added to the HTTP header as Authorization.}
\usage{
create_auth(request_method, headers, payload, uri, query, date_stamp,
  region_name, service_name, request_time = now_s3())
}
\arguments{
\item{request_method}{the HTTP verb used for the request (GET/POST/PUT etc)}

\item{headers}{list of headers to be included in the request. Must include a
\code{host} header. See examples for correct format}

\item{payload}{the payload from the body of the HTTP/HTTPS request}

\item{query}{the query parameters of the request, url encoded}

\item{date_stamp}{request date in the form YYYYMMDD; defaults to the current
date.}

\item{region_name}{name of the AWS region being targeted, e.g. 'eu-west-1'}

\item{service_name}{name of the AWS service being targeted, e.g. 'ec2'}

\item{request_time}{string containing the date and time of the request,
matching the value used in previous steps, in the form YYYYMMDDTHHMMSSZ}

\item{endpoint}{the AWS endpoint to send the request to}
}
\value{
authorisation string to be included in request headers
}
\description{
This is the final step in the authorization procedure, where the three tasks
are put together to create the authorization value.
}

