% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create_credential_scope}
\alias{create_credential_scope}
\title{Create the credential scope string.}
\usage{
create_credential_scope(date_stamp = date_stamp, region, service)
}
\arguments{
\item{date_stamp}{date in the form YYYYMMDD - must match that used in other
steps}

\item{region}{region being targeted}

\item{service}{being targeted}
}
\description{
Helper function for concatenating strings into the right format for the
credential scope value.
}
\examples{
credential_scope('20110909', 'us-east-1', 'iam')
}

